Add method to perform long polling on the model
"receiveMessages" sends a request to the server to get the messages; if
there is none to got the server keeps the connection open, causing
"receiveMessages" to wait until some message is available (or the
timeout ended, in which case the request would succeed but no message
would be returned). Once a response is got, "receiveMessages" sends a
new request to the server, and the process is repeated again and again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>