Use a newer version of the nextcloudci/php7.1 Docker image in Drone
The "php7.1" and "php7.1-integration" tests are not actually being run
in Drone; when the command "more ./tests/autoconfig-oracle.php" is
executed by "core_install.sh" the execution of the remaining Drone
commands is broken (strange... probably some black magic related to TTYs
or something like that), and thus the tests are not run (although no
error is issued, so it wrongly seems that they were successful).

The problem occurs on CentOS based Docker images, but not on Debian
based ones. Therefore, the base images for the "php7.1" and
"php7.1-integration" tests were changed to use a newer version of the
"nextcloudci/php7.1" image, which is now based on Debian instead of
CentOS like the older ones.

This also removes the need of installing "wget", as it is already
included in the newer image too.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>