Fix Strict Standards error in integration tests on PHP 7
Besides returning the last element, "end" also sets the internal pointer
of the array to the last element, and for that it receives a reference
to the array. Thus, "end(array_keys(..." can not be used, as it would
generate the Strict Standards error "Only variables should be passed by
reference", so the returned value from "array_keys" has to be stored in
a variable.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>