Generalize ParticipantView CSS rules
Instead of being based on the "participantTabView" id the CSS rules used
by the ParticipantView were modified to be based on its class,
"participantWithList". This will make possible to change the parent
element of the ParticipantView and keep its style.

The rules for links were merged as those links that required a padding
due to being shown with an icon were also those shown inside the list
items.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>