Make ParticipantCollection to automatically fetch when its room changes
The ParticipantCollection provides the list of participants (including
guests) in a room. The Room model is synced through signaling and
provides some information about the registered users and the guests in a
room, so the ParticipantCollection can listen to changes in those
properties to automatically fetch its items again when needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>