Add a Backbone model for the local storage of the browser
The LocalStorageModel makes possible to use the local storage of the
browser as a Backbone model, for example, with a Marionette view.

It is introduced to be used in a later commit as a bridge between the
local storage of the browser and an EditableTextLabel for the name of
the guest.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>