Replace custom code with LocalStorageModel for guest name
Only the code related to getting and setting the guest name in the UI
was modified to use the LocalStorageModel; direct calls to localStorage
to simply get the value in other areas were kept (and this change does
not affect their behaviour).

As "maxlength" in the guest name input field is set to "20" there is no
need to check the length of the new nick before setting it; if some
validation of the nick was needed before setting it it would have to be
done by using "validate" from Backbone on the model (and
EditableTextLabel would have to be extended too to provide feedback to
the user in that case).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>