Fix sidebar in fullscreen mode
In normal mode (and a wide screen), when the sidebar is opened a
"margin-right" is applied to the app content to make room for the
sidebar. In fullscreen mode the element set as fullscreen is the app
content, but it seems that different browsers handle fullscreen elements
with a margin in different ways: Firefox ignores the margin and uses the
full width, while Chromium uses a strange mix which is neither the full
width nor the full margin. Due to this, now the margin is removed in
fullscreen mode to unify the appearance across browsers.

As there is no margin, the sidebar now visually seems to slide on the
app content instead of "pushing" it to the left (which is the same
behaviour shown in normal mode and narrow screens). Fortunately, this
behaviour fits well with the fullscreen mode, as the video is arguably
the reason to enter in fullscreen mode while the sidebar in this case is
probably just a temporal view that is closed most of the time.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>