Show link icon next to room name when the user can not moderate
The main room name element is now shown as "inline-block", which causes
the link icon to be placed next to it instead of below.

Due to the change from "block" to "inline-block" there is less available
width for the children input fields; the previous maximum width value is
now too narrow, so that limit to the width was removed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>