Add the message to the mention notifications
The chat message can be much longer than the maximum length allowed for
notification messages, so the notification message is trimmed if needed.

When the notification is being prepared to be displayed it is taken into
account if the message was trimmed, and in that case an ellipsis is
added at the start and/or the end (depending on the case) of the
message.

Note that the original notification message does not contain the
ellipsis characters, as depending on the language a whitespace should be
added or not before and after an ellipsis, and the language in which the
notification will be displayed is only known at the time of displaying
it, but not when it is saved.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>