Fix integration tests for actions that just require being in a room
Since commit eae71a3977068b578c590982c3b9ebacf9f23074 a participant
(user or guest) can be in a room without joining a call; since then
sessions are no longer set and passwords are no longer checked when
joining a call, but when joining a room. Due to this some integration
tests had to be modified to join a room instead of a call.

Note, however, that in the modified tests there is no need to join a
call too once the participant has joined the room, as they either test
joining the room itself (the password tests) or the tested actions
simply need the participant to be in a room (the chat tests). Moreover,
in this later case, note that a participant is in a room if she created
a room, was invited to it or explicitly joined it, so the chat tests
involving a participant creating a room or inviting another one to it
did not need to be modified.

Finally, neither the call tests nor the "userJoinsCall" method from
FeatureContext were modified yet, as it has to be decided first whether
directly joining a call also joins the room automatically or not.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>