Fix SQLite issues where you can only connect to the newest room
The problem is, that SQLite does the LIMIT before the LEFT JOIN.
This means it would always take the latest room and then try to
match your session against it. However, we wanted to first match
the LEFT JOIN and then limit the result set to 1 row.

By changing the FROM to the participant table, we will always match
the correct participant (if there is one) and then left join the
room just like we intended to.

Signed-off-by: Joas Schilling <coding@schilljs.com>