Set height for header element in public pages
The header element contains the header div; as the header div uses a
fixed position the header element has no height, so it must be
explicitly set to prevent #app-content-wrapper from overlapping with the
header element.

Right now it was not very noticeable, as it only affected the empty
content; as it name implies most of the space was empty except for a
message at the centre of the screen, so visually it did not overlap with
the header. However having a proper height set for the header will be a
must once the chat view is shown in the app content wrapper.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>