Show the chat view as the main view when in a room but not in a call
When shown as the main view, the input field to add a new message is
always shown and a scroll bar is provided just for the list of messages.

Only the chat view is added and removed to and from the main view; the
other elements in the main view are not modified when that happens, and
they are hidden using some CSS magic.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>