Fix sorting of chat messages with same timestamp
The timestamp is in seconds, so when sent extremely fast two or more
messages can have the same timestamp. As a stable sorting algorithm is
used and the messages were sorted only by timestamp, when two or more
messages had the same timestamp the relative order between them was the
same as the one returned by the server. Due to this those messages were
sorted from newest to oldest between them, while the rest were sorted
from oldest to newest. The solution is to sort the models by their
reversed position in the set returned by the server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>