Add ElementExpanderButton to tabs in the sidebar
Due to the scroll bar being shown only for the tabs the call details are
always visible. In short screens this may not leave too much space for
the tabs, making them uncomfortable to use. To improve that now an
ElementExpanderButton is added automatically to every tab added to the
sidebar; clicking on the button toggles the associated tab between its
normal size and an expanded size that fills the whole sidebar.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>