Fix postgres issue when updating
Postgres could not be used in 12. It installed the tables with wrong names
so we need to ignore it on the update steps. The wrong tables will be
dropped and new tables will be created with further migrations.
But we shouldn't try to migrate data from the old and broken layout.

Signed-off-by: Joas Schilling <coding@schilljs.com>