Allow postgres for update but not < 2.0.0
Actually it's not postgres having the problem, but migrations.
Postgres allows camelCased columns, but they need to be quoted.
They were quoted from the old database.xml file but this does
not work anymore in migrations. But that means it worked on
postgres in the past, it just didn't work in 2.9~ to install
since migrations where merged. So we are reverting the check
for postgres and only allow updating from 2.0.0 or later.

This makes sure that when newly installing the migration code
is not run.

Signed-off-by: Joas Schilling <coding@schilljs.com>