Make sure Postgres uses lowercase for the migration generated columns
The issue is that migrations always generate lowercase columns only.
So all columns that were created with database.xml are okay and cC.
But all intermediate columns the migrations create need to be
referenced as lowercase columns in the queries.

In the future we won't have this problem anymore because we switched
to under_score column names for 3.0

Signed-off-by: Joas Schilling <coding@schilljs.com>