Show date in first message
When using the oldest to newest layout now the first (oldest) message
always shows the date.

When using the newest to oldest layout, however, there is no date shown
in the first (newest) message; in this case it would be good to show it
only if the newest message was not sent today, but the newest to oldest
layout is currently not used anywhere (and it is likely to be removed
from the code), so there was no need to spend time on that.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>