Add base configuration for acceptance tests
This commits adds the base configuration needed to run acceptance tests
for the Talk app.

The script to run the acceptance tests simply calls the "run.sh" script
from the server acceptance tests specifying the directory of the
acceptance tests for the Talk app.

The installation script called before the acceptance tests are actually
run is extended from the one used in the acceptance tests for the server
to also enable the Talk app.

Due to Behat limitations, the Behat configuration used in the acceptance
tests for the server can not be extended; the configuration used is
exclusive for the Talk app. Due to that some parameters must be
duplicated from the Behat configuration for the server acceptance tests
(like the session definitions).

In any case, the core library and all the contexts for the server are
automatically available in the acceptance tests for apps; their path
does not need to be set in the "autoload" section, although the specific
contexts to be used have to be explicitly set; the ones included are
those needed (either directly or due to dependencies) to be able to use
the "I am logged in"/"I am logged in as the admin" steps, which are the
most basic steps needed by the acceptance tests for almost every app.

On the other hand, the paths to the features and the contexts from the
Talk app must be explicitly set in the configuration.

Finally, note that this commit does not include any test feature, so
calling "run.sh" would just end in a failure; some basic test features
will be added in a following commit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>