Run the acceptance tests for conversations in Drone
The acceptance tests requires the grandparent directory of the root
directory of the app to be the root directory of the Nextcloud server.
Drone only clones the app repository, so the Nextcloud server repository
has to be explicitly cloned and the app moved to the "apps" directory
before the tests are run.

Besides that, as the acceptance tests need to be run on Apache,
"/var/www/html" has to be linked to the root directory of the Nextcloud
server.

Once all that is done the tests can be run in the same way that they are
run in Drone for the Nextcloud server but just specifying the acceptance
tests directory too.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>