Show display name instead of user name in formatted mentions
In some cases (like when LDAP is being used) a user can not be easily
identified through its user name (as it is just a hash), so formatted
mentions now show the display name in all cases (and if the server does
not provide a display name the "name" parameter of the template falls
back to the user name).

Note that the acceptance tests do not need to be updated, as the
display name of the mentioned user is the same as its user name.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>