Show contacts menu when clicking on a mention
The contacts menu expects its parent element to be its positioning
context, so the mention must have a relative position. Besides that, the
default CSS rules from the server set the position of the contacts menu
assuming that the parent element is an avatar, so the rules have to be
overriden to position it on the text.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>