Fix appearance of author row when showing the date
When a message shows the date separator its author row is explicitly
shown. This is needed to ensure that the author row will always be shown
in that case, even if the message is a grouped message. In normal
messages (those that are not grouped and do not show the date separator)
the author row is shown using "display: inline-flex", so the same value
should be used too for messages that show the date separator.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>