Remove support for newest on top chat layout
When the oldest on top layout was added it was enabled by default,
although the original newest on top layout was also kept as an option.
However, the newest on top layout has not been used since then, and it
is not expected to be used ever again. To make the chat view more
maintainable and ease adding new features this commit removes the
original newest on top chat layout.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>