Do not take margins into account for the container scroll height
The scroll area covers the top and bottom padding of an element, but not
its top and bottom margins, so the margins should not be taken into
account in the scroll calculations.

Currently this does not cause any difference in the behaviour, as the
message lists does not have any margin, but it is fixed just in case it
is added in the future.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>