Make acceptance test for two users chatting more robust
The chat between two users was tested by a sequence of one user sending
a message and immediately after that the other user sending another
message. Most of the time the messages were processed in the server in
the order in which they were sent, but sometimes it could happen that a
previous message took longer than usual to get to the server and then
the next message from the other user was processed before it; in those
cases the test failed due to the messages not matching the expected
order. Now each user waits for the last message of the other user to be
received before sending the next message.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>