Force inactive users to leave the room
Inactive guests (those that have not sent a ping in the last 30 seconds)
are automatically cleaned from a room when the information for that room
is got. In the same way, now inactive users are forced to leave the room
too when the information for the room is got.

Although all inactive guests are cleaned all at once inactive users are
forced to leave the room one by one (each one with its own SQL query);
although it would be better to force all the inactive users to leave the
room at once this should not be a big problem, as users would usually
leave the room cleanly; only crashed clients or other exceptional
situations should cause a user to be inactive.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>